Add test for git-submodule Perl error
Which ensures that the following error/warning isn't printed on get:

    Cloning into '/tmp/build/get'...
    Fetching HEAD
    d72625f Merge pull request #259 from alphagov/upgrade-cf-118272895
    /usr/libexec/git-core/git-submodule: line 1: /usr/bin/perl: not found
    /usr/libexec/git-core/git-submodule: line 1: /usr/bin/perl: not found

This occurred for us because we forked the resource and tried a different
base image. The tests still pass and the resource still works, but you get
that distracting message on every get.

It happens because git-submodule depends on Perl to do some pattern
matching. This doesn't normally affect the resource because Perl was added
to the base image which this container builds from in:

- concourse/core-busyboxplus@895aad2

Adding a test to the resource itself will prevent a similar regression in
the future, even when using the intended base image.

Matching substrings in sh (not bash) is pretty limited, so I'm piping to
grep and inverting the exit code. It will also print the error message when
the test fails. I'm not matching the exact path of `perl` because it may
change.