Add Perl to packages installed
This makes our modified container pass the new test that was added in the
preceding commit and upstream as concourse/git-resource#/58.

Failure from before looks like this:

    running it_can_use_submodlues_without_perl_warning...
      Switched to a new branch 'bogus'
      Switched to branch 'master'
      Switched to a new branch 'bogus'
      Switched to branch 'master'
      Cloning into 'repo.eOMCDN'...
      /usr/libexec/git-core/git-submodule: line 1: /usr/bin/perl: not found
      /usr/libexec/git-core/git-submodule: line 1: /usr/bin/perl: not found
    The command '/bin/sh -c /opt/resource-tests/all.sh &&   rm -rf /tmp/*' returned a non-zero code: 1

It increases the size of the resulting image by this much:

    ➜  paas-git-resource git:(fix_submodule_perl) docker images | awk '$1 == "git-resource"'
    git-resource                         fix_submodule_perl   6c682b6e425f        19 minutes ago      88.76 MB
    git-resource                         gds_master           5db676ae6bdc        20 minutes ago      47.9 MB

This isn't necessarily a permanent solution as we may decide to go back to
the busyboxplus image when we submit our GPG changes upstream. See Hector's
description in:

- https://www.pivotaltracker.com/story/show/118562893