Prevent check returning older commits
There are occasions where check can return commits older than the
requested ref that it hasn't previously returned. This can cause
concourse to run with these older commits becauseof how it updates its
internal list of versions. This issue is most commonly observed with a
freshly pushed pipeline.

For example given a pipeline pointed at a git branch as below with HEAD
pointing at E:
```
A--B------E
    \    /
     C--D
```
Concourse will initially call check with no `ref`. It will then
correctly return E. Next time concourse polls, it will call check with a
`ref` of E. In this case check will return C,D,E (because it does a `git
log E~1..HEAD`). At this point, concourse inserts C and D into its
database, and because they're more recently added than E, they become
the latest versions in its view, and therefore the pipeline will run
with commit D.

This fixes this situation by testing whether the ref commit is the HEAD
of the repo, and in that case it falls back to returning the HEAD commit
only to avoid this problem.